requirements.txt
graphene
asgiref==3.2.3
backcall==0.1.0
boto3==1.11.6
botocore==1.14.6
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
Django==3.0.2
django-braces==1.14.0
django-crispy-forms==1.8.1
django-extensions==2.2.6
django-rest-framework==0.1.0
django-warrant==0.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
ecdsa==0.15
envs==1.3
factory-boy==2.12.0
Faker==4.0.0
future==0.18.2
idna==2.8
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
jmespath==0.9.4
mock==3.0.5
mysqlclient==1.4.6
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
ptyprocess==0.6.0
Pygments==2.5.2
PyJWT==1.7.1
python-dateutil==2.8.1
python-jose-cryptodome==1.3.2
pytz==2019.3
requests==2.22.0
s3transfer==0.3.1
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.7
warrant==0.6.1
wcwidth==0.1.8
whitenoise==5.0.1
graphene-django
